Refactoring external data location coordinates #3
